Remove FileUploader theme tokens
