+++ packages/react/src/primitives/shared/types.ts
@@ -16,7 +16,6 @@ type ComponentNameKey =
   | 'DropZone'
   | 'Expander'
   | 'Fieldset'
-  | 'FileUploader'
   | 'Flex'
   | 'Grid'
   | 'Heading'
@@ -117,24 +116,6 @@ type ComponentClassNameKey =
   | 'FieldGroupFieldWrapper'
   | 'Fieldset'
   | 'FieldsetLegend'
-  | 'FileUploaderDropZone'
-  | 'FileUploaderDropZoneIcon'
-  | 'FileUploaderDropZoneText'
-  | 'FileUploaderDropZoneButton'
-  | 'FileUploaderFile'
-  | 'FileUploaderFileWrapper'
-  | 'FileUploaderFileName'
-  | 'FileUploaderFileSize'
-  | 'FileUploaderFileInfo'
-  | 'FileUploaderFileStatus'
-  | 'FileUploaderFileImage'
-  | 'FileUploaderFileMain'
-  | 'FileUploaderLoader'
-  | 'FileUploaderPreviewer'
-  | 'FileUploaderPreviewerText'
-  | 'FileUploaderPreviewerBody'
-  | 'FileUploaderPreviewerFooter'
-  | 'FileUploaderPreviewerFooterActions'
   | 'Flex'
   | 'Grid'
   | 'Heading'
