+++ packages/ui/src/theme/__tests__/defaultTheme.test.ts
@@ -696,70 +696,6 @@ describe('@aws-amplify/ui', () => {
         --amplify-components-fieldset-outlined-large-padding: var(--amplify-space-large);
         --amplify-components-fieldset-small-gap: var(--amplify-components-field-small-gap);
         --amplify-components-fieldset-large-gap: var(--amplify-components-field-large-gap);
-        --amplify-components-fileuploader-dropzone-background-color: var(--amplify-colors-background-primary);
-        --amplify-components-fileuploader-dropzone-border-radius: var(--amplify-radii-small);
-        --amplify-components-fileuploader-dropzone-border-color: var(--amplify-colors-border-primary);
-        --amplify-components-fileuploader-dropzone-border-style: dashed;
-        --amplify-components-fileuploader-dropzone-border-width: var(--amplify-border-widths-small);
-        --amplify-components-fileuploader-dropzone-gap: var(--amplify-space-small);
-        --amplify-components-fileuploader-dropzone-padding-block: var(--amplify-space-xl);
-        --amplify-components-fileuploader-dropzone-padding-inline: var(--amplify-space-large);
-        --amplify-components-fileuploader-dropzone-text-align: center;
-        --amplify-components-fileuploader-dropzone-active-background-color: var(--amplify-colors-brand-primary-10);
-        --amplify-components-fileuploader-dropzone-active-border-radius: var(--amplify-components-fileuploader-dropzone-border-radius);
-        --amplify-components-fileuploader-dropzone-active-border-color: var(--amplify-colors-border-pressed);
-        --amplify-components-fileuploader-dropzone-active-border-style: var(--amplify-components-fileuploader-dropzone-border-style);
-        --amplify-components-fileuploader-dropzone-active-border-width: var(--amplify-border-widths-medium);
-        --amplify-components-fileuploader-dropzone-icon-color: var(--amplify-colors-border-primary);
-        --amplify-components-fileuploader-dropzone-icon-font-size: var(--amplify-font-sizes-xxl);
-        --amplify-components-fileuploader-dropzone-text-color: var(--amplify-colors-font-tertiary);
-        --amplify-components-fileuploader-dropzone-text-font-size: var(--amplify-font-sizes-medium);
-        --amplify-components-fileuploader-dropzone-text-font-weight: var(--amplify-font-weights-bold);
-        --amplify-components-fileuploader-file-background-color: var(--amplify-colors-background-primary);
-        --amplify-components-fileuploader-file-border-radius: var(--amplify-radii-small);
-        --amplify-components-fileuploader-file-border-color: var(--amplify-colors-border-primary);
-        --amplify-components-fileuploader-file-border-style: solid;
-        --amplify-components-fileuploader-file-border-width: var(--amplify-border-widths-small);
-        --amplify-components-fileuploader-file-padding-block: var(--amplify-space-xs);
-        --amplify-components-fileuploader-file-padding-inline: var(--amplify-space-small);
-        --amplify-components-fileuploader-file-gap: var(--amplify-space-small);
-        --amplify-components-fileuploader-file-align-items: baseline;
-        --amplify-components-fileuploader-file-name-font-size: var(--amplify-font-sizes-medium);
-        --amplify-components-fileuploader-file-name-font-weight: var(--amplify-font-weights-bold);
-        --amplify-components-fileuploader-file-name-color: var(--amplify-colors-font-primary);
-        --amplify-components-fileuploader-file-size-font-size: var(--amplify-font-sizes-small);
-        --amplify-components-fileuploader-file-size-font-weight: var(--amplify-font-weights-normal);
-        --amplify-components-fileuploader-file-size-color: var(--amplify-colors-font-tertiary);
-        --amplify-components-fileuploader-file-image-width: var(--amplify-space-xxl);
-        --amplify-components-fileuploader-file-image-height: var(--amplify-space-xxl);
-        --amplify-components-fileuploader-file-image-background-color: var(--amplify-colors-background-secondary);
-        --amplify-components-fileuploader-file-image-color: var(--amplify-colors-font-tertiary);
-        --amplify-components-fileuploader-file-image-border-radius: var(--amplify-radii-small);
-        --amplify-components-fileuploader-loader-stroke-linecap: round;
-        --amplify-components-fileuploader-loader-stroke-empty: var(--amplify-colors-border-secondary);
-        --amplify-components-fileuploader-loader-stroke-filled: var(--amplify-components-loader-stroke-filled);
-        --amplify-components-fileuploader-loader-stroke-width: var(--amplify-border-widths-large);
-        --amplify-components-fileuploader-previewer-background-color: var(--amplify-colors-background-primary);
-        --amplify-components-fileuploader-previewer-border-color: var(--amplify-colors-border-primary);
-        --amplify-components-fileuploader-previewer-border-style: solid;
-        --amplify-components-fileuploader-previewer-border-width: var(--amplify-border-widths-small);
-        --amplify-components-fileuploader-previewer-border-radius: var(--amplify-radii-small);
-        --amplify-components-fileuploader-previewer-padding-block: var(--amplify-space-zero);
-        --amplify-components-fileuploader-previewer-padding-inline: var(--amplify-space-zero);
-        --amplify-components-fileuploader-previewer-max-height: 40rem;
-        --amplify-components-fileuploader-previewer-max-width: auto;
-        --amplify-components-fileuploader-previewer-text-font-size: var(--amplify-font-sizes-medium);
-        --amplify-components-fileuploader-previewer-text-font-weight: var(--amplify-font-weights-bold);
-        --amplify-components-fileuploader-previewer-text-color: var(--amplify-colors-font-primary);
-        --amplify-components-fileuploader-previewer-body-padding-block: var(--amplify-space-medium);
-        --amplify-components-fileuploader-previewer-body-padding-inline: var(--amplify-space-medium);
-        --amplify-components-fileuploader-previewer-body-gap: var(--amplify-space-small);
-        --amplify-components-fileuploader-previewer-footer-border-color: var(--amplify-colors-border-secondary);
-        --amplify-components-fileuploader-previewer-footer-border-style: solid;
-        --amplify-components-fileuploader-previewer-footer-border-width: var(--amplify-border-widths-small);
-        --amplify-components-fileuploader-previewer-footer-padding-block: var(--amplify-space-medium);
-        --amplify-components-fileuploader-previewer-footer-padding-inline: var(--amplify-space-medium);
-        --amplify-components-fileuploader-previewer-footer-justify-content: space-between;
         --amplify-components-flex-gap: var(--amplify-space-medium);
         --amplify-components-flex-justify-content: normal;
         --amplify-components-flex-align-items: stretch;
@@ -1101,9 +1037,9 @@ describe('@aws-amplify/ui', () => {
         --amplify-components-storagemanager-dropzone-padding-inline: var(--amplify-space-large);
         --amplify-components-storagemanager-dropzone-text-align: center;
         --amplify-components-storagemanager-dropzone-active-background-color: var(--amplify-colors-brand-primary-10);
-        --amplify-components-storagemanager-dropzone-active-border-radius: var(--amplify-components-fileuploader-dropzone-border-radius);
+        --amplify-components-storagemanager-dropzone-active-border-radius: var(--amplify-components-storagemanager-dropzone-border-radius);
         --amplify-components-storagemanager-dropzone-active-border-color: var(--amplify-colors-border-pressed);
-        --amplify-components-storagemanager-dropzone-active-border-style: var(--amplify-components-fileuploader-dropzone-border-style);
+        --amplify-components-storagemanager-dropzone-active-border-style: var(--amplify-components-storagemanager-dropzone-border-style);
         --amplify-components-storagemanager-dropzone-active-border-width: var(--amplify-border-widths-medium);
         --amplify-components-storagemanager-dropzone-icon-color: var(--amplify-colors-border-primary);
         --amplify-components-storagemanager-dropzone-icon-font-size: var(--amplify-font-sizes-xxl);
@@ -1130,6 +1066,8 @@ describe('@aws-amplify/ui', () => {
         --amplify-components-storagemanager-file-image-background-color: var(--amplify-colors-background-secondary);
         --amplify-components-storagemanager-file-image-color: var(--amplify-colors-font-tertiary);
         --amplify-components-storagemanager-file-image-border-radius: var(--amplify-radii-small);
+        --amplify-components-storagemanager-filelist-flex-direction: column;
+        --amplify-components-storagemanager-filelist-gap: var(--amplify-space-small);
         --amplify-components-storagemanager-loader-stroke-linecap: round;
         --amplify-components-storagemanager-loader-stroke-empty: var(--amplify-colors-border-secondary);
         --amplify-components-storagemanager-loader-stroke-filled: var(--amplify-components-loader-stroke-filled);
@@ -1150,8 +1088,6 @@ describe('@aws-amplify/ui', () => {
         --amplify-components-storagemanager-previewer-body-padding-inline: var(--amplify-space-medium);
         --amplify-components-storagemanager-previewer-body-gap: var(--amplify-space-small);
         --amplify-components-storagemanager-previewer-footer-justify-content: flex-end;
-        --amplify-components-storagemanager-filelist-flex-direction: column;
-        --amplify-components-storagemanager-filelist-gap: var(--amplify-space-small);
         --amplify-components-switchfield-disabled-opacity: var(--amplify-opacities-60);
         --amplify-components-switchfield-focused-shadow: 0px 0px 0px 2px var(--amplify-colors-border-focus);
         --amplify-components-switchfield-font-size: var(--amplify-font-sizes-medium);
