+++ packages/ui/src/theme/__tests__/themeType.test.ts
@@ -23,7 +23,6 @@ describe('Amplify UI Theme', () => {
             fieldcontrol: {},
             fieldgroup: {},
             fieldmessages: {},
-            fileuploader: {},
             flex: {},
             heading: {},
             highlightmatch: {},
@@ -79,7 +78,6 @@ describe('Amplify UI Theme', () => {
               "fieldcontrol": Object {},
               "fieldgroup": Object {},
               "fieldmessages": Object {},
-              "fileuploader": Object {},
               "flex": Object {},
               "heading": Object {},
               "highlightmatch": Object {},
