+++ packages/ui/src/types/primitives/componentClassName.ts
@@ -61,24 +61,6 @@ export enum ComponentClassName {
   FieldGroupFieldWrapper = 'amplify-field-group__field-wrapper',
   Fieldset = 'amplify-fieldset',
   FieldsetLegend = 'amplify-fieldset__legend',
-  FileUploaderDropZone = 'amplify-fileuploader__dropzone',
-  FileUploaderDropZoneIcon = 'amplify-fileuploader__dropzone__icon',
-  FileUploaderDropZoneText = 'amplify-fileuploader__dropzone__text',
-  FileUploaderDropZoneButton = 'amplify-fileuploader__dropzone__button',
-  FileUploaderFile = 'amplify-fileuploader__file',
-  FileUploaderFileWrapper = 'amplify-fileuploader__file__wrapper',
-  FileUploaderFileName = 'amplify-fileuploader__file__name',
-  FileUploaderFileSize = 'amplify-fileuploader__file__size',
-  FileUploaderFileInfo = 'amplify-fileuploader__file__info',
-  FileUploaderFileImage = 'amplify-fileuploader__file__image',
-  FileUploaderFileMain = 'amplify-fileuploader__file__main',
-  FileUploaderFileStatus = 'amplify-fileuploader__file__status',
-  FileUploaderLoader = 'amplify-fileuploader__loader',
-  FileUploaderPreviewer = 'amplify-fileuploader__previewer',
-  FileUploaderPreviewerText = 'amplify-fileuploader__previewer__text',
-  FileUploaderPreviewerBody = 'amplify-fileuploader__previewer__body',
-  FileUploaderPreviewerFooter = 'amplify-fileuploader__previewer__footer',
-  FileUploaderPreviewerFooterActions = 'amplify-fileuploader__previewer__footer__actions',
   Flex = 'amplify-flex',
   Grid = 'amplify-grid',
   Heading = 'amplify-heading',
